third_party/blink/web_tests/fast/selectors/query-update-distribution-expected.txt

CONSOLE WARNING: line 15: Element.createShadowRoot is deprecated and will be removed in M80, around February 2020. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
Should update distribution when needed for querySelector and related methods.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS aRoot.querySelector(':host-context(#c) #d') is d
PASS toArray(aRoot.querySelectorAll(':host-context(#c) #d')) is [d]
PASS hostRoot.querySelector('::content #a') is null
PASS toArray(hostRoot.querySelectorAll('::content #a')) is []
PASS a.matches('::content #a') is false
PASS d.matches(':host-context(#host) #d') is true
PASS d.matches(':host-context(#c) #d') is true
PASS b.closest('::content #a') is null
PASS e.closest(':host-context(#host) #d') is d
PASS successfullyParsed is true

TEST COMPLETE
